agent_service/requirements.txt
langchain
langchain_ollama
langgraph
langchain-community
python-dotenv
requests
langchain-openai
fastapi
uvicorn
langsmith
langchain-experimental